requirements.txt
Flask==2.2.3
google-generativeai
gunicorn==20.0.4
grpcio
Werkzeug==2.2.3
langchain
langchain_community
sentence-transformers
faiss-cpu
